refactor make_index to be simpler

+++ shotglass/app/management/commands/make_index.py
@@ -10,6 +10,7 @@ import logging
 import os
 import re
 import subprocess
+import sys
 
 import django.db
 from django.core.management.base import BaseCommand
@@ -22,25 +23,12 @@ from app import render
 INDEX_SUFFIXES = ('.c', '.py')
 BORING_DIRS = ('.pc',) # TODO: make configurable
 
-if 01:
-    # XX: never flushes logs
-    logger = logging.getLogger(__name__)
-else:
-    class HackLogger(object):
-        def logme(self, myformat, *args):
-            print '*', myformat % args
-        debug = info = warning = logme
-    logger = HackLogger()
-
-
-# def calc_radon(path):
-#     # return iter of blocks
-#     code = open(path).read()
-#     try:
-#         return cc_visit(code)
-#     except SyntaxError:
-#         # logger.warning('%s: %s', path, error)
-#         return []
+logging.basicConfig(
+    stream=sys.stderr,
+    level=logging.DEBUG,
+    format='%(levelname)s %(module)s %(message)s',
+)
+logger = logging.getLogger(__name__)
 
 
 def format_project_name(project_dir):
@@ -48,10 +36,6 @@ def format_project_name(project_dir):
 
 def test_format_project_name():
     assert format_project_name('/a/b/') == 'b'
-
-# def strip_project_dir(project_dir, path):
-#     return path[len(project_dir):].lstrip('/')
-
 
 def walk_type(topdir, name_func):
     for root, dirs, names in os.walk(topdir, topdown=True):
@@ -62,33 +46,27 @@ def walk_type(topdir, name_func):
         for path in paths:
             yield path
 
-
 def make_index(project, project_dir):
-#     is_c = re.compile(r'\.c$').search
-#     is_python = re.compile(r'\.py$').search
+    is_python = re.compile(r'\.py$').search
     
     if django.db.connection.vendor == 'sqlite':
         django.db.connection.cursor().execute('PRAGMA synchronous=OFF')
 
     # XX: delete project's index
     # pylint: disable=no-member
-    proj_symbols = models.SourceLine.objects.filter(project=project)
-    proj_symbols.delete()
-    
-#     c_paths = list(walk_type(project_dir, is_c))
-#     logger.info('%s: %d C files', project, len(c_paths))
-
-#     index_c_mccabe(project, project_dir, c_paths)
-#     # shows as "0" because of the PRAGMA SYNC above
-
+    logger.info("project %s", project)
+    if 1:
+        logger.info("%s: zapping old data", project)
+        proj_symbols = models.SourceLine.objects.filter(project=project)
+        proj_symbols.delete()
+ 
     py_paths = list(walk_type(project_dir, is_python))
     logger.info('%s: %d Python files', project, len(py_paths))
-#     index_py_radon(project, project_dir, py_paths)
 
 #     render.render(proj_symbols)
 
 class Command(BaseCommand):
-    help = 'beer'
+    help = __doc__
 
     def add_arguments(self, parser):
         parser.add_argument('project_dirs', metavar='FILE', nargs='+')
